02_backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   starlette
    #   watchfiles
argon2-cffi==25.1.0
    # via minio
argon2-cffi-bindings==25.1.0
    # via argon2-cffi
async-timeout==5.0.1
    # via redis
cachetools==7.0.6
    # via pyld
certifi==2024.2.2
    # via
    #   minio
    #   requests
cffi==2.0.0
    # via
    #   argon2-cffi-bindings
    #   cryptography
charset-normalizer==3.4.4
    # via requests
click==8.1.7
    # via
    #   typer
    #   uvicorn
cryptography==45.0.7
    # via -r requirements.in
dnspython==2.6.1
    # via -r requirements.in
exceptiongroup==1.2.0
    # via anyio
fastapi==0.110.1
    # via -r requirements.in
frozendict==2.4.7
    # via pyld
greenlet==3.0.3
    # via sqlalchemy
h11==0.14.0
    # via uvicorn
httptools==0.6.1
    # via uvicorn
idna==3.6
    # via
    #   anyio
    #   requests
isodate==0.7.2
    # via rdflib
lxml==6.1.0
    # via pyld
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
minio==7.2.18
    # via -r requirements.in
psycopg2-binary==2.9.9
    # via -r requirements.in
pycparser==2.23
    # via cffi
pycryptodome==3.23.0
    # via minio
pydantic==2.6.4
    # via fastapi
pydantic-core==2.16.3
    # via pydantic
pygments==2.18.0
    # via rich
pyld==3.0.0
    # via -r requirements.in
pyparsing==3.3.2
    # via rdflib
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   uvicorn
pyyaml==6.0.1
    # via
    #   -r requirements.in
    #   uvicorn
rdflib==7.6.0
    # via -r requirements.in
redis==7.0.0
    # via -r requirements.in
requests==2.32.5
    # via -r requirements.in
rich==13.7.1
    # via
    #   -r requirements.in
    #   typer
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.29
    # via -r requirements.in
starlette==0.37.2
    # via fastapi
typer==0.12.3
    # via -r requirements.in
typing-extensions==4.11.0
    # via
    #   anyio
    #   fastapi
    #   minio
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typer
    #   uvicorn
urllib3==2.5.0
    # via
    #   minio
    #   requests
uvicorn[standard]==0.29.0
    # via -r requirements.in
uvloop==0.21.0
    # via uvicorn
watchfiles==0.21.0
    # via uvicorn
websockets==12.0
    # via uvicorn
